requirements.txt.txt
tensorflow
streamlit
gdown
numpy
Pillow
